[feat]コメント追加/propsをtsxに移動 #14
